Assignment 1

+++ group_project/assignment_1.py
@@ -1,6 +1,6 @@
 def first_half(a_string):
     if len(a_string)%2==0:
-        return a_string[0:len(a_string) / 2]
+        return a_string[0:int(len(a_string) / 2)]
     if len(a_string)%2==1:
         return a_string[0:(len(a_string)/2)+1]
